查看是否加入趣吧返回趣吧id

+++ src/main/java/com/ruoyi/project/system/quba/mapper/QubaMapper.java
@@ -33,7 +33,7 @@ public interface QubaMapper {
 
     List<Quba> getByUserId(Long userId);
 
-    int getUserJoined(Long userId);
+    Long getUserJoined(Long userId);
 
     Long getUserJoinedQuba(Long userId);
